requirements.txt
alembic==1.13.2
Flask==3.0.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Jinja2==3.1.4
psycopg2-binary==2.9.9
python-dotenv==1.0.1
SQLAlchemy==2.0.31
